Tests fixed and JUnit test added for #5311
svn changeset:14100/svn branch:6.4

+++ tests/src/com/vaadin/tests/dd/AcceptAnythingWindow.java
@@ -76,9 +76,9 @@ public class AcceptAnythingWindow extends Window {
 
                         try {
                             layout.getPosition(component).setTopValue(
-                                    clientY - top);
+                                    Float.valueOf(clientY - top));
                             layout.getPosition(component).setLeftValue(
-                                    clientX - left);
+                                    Float.valueOf(clientX - left));
                         } catch (Exception e) {
                             // TODO: handle exception
                         }
